Improve the performance of indirect string equality.
We define that all strings in the database are unique. That means that
if both strings are in the store then we can  simply compare their
addresses. The corresponding in-memory case isn't quite so simple: we
can't immediately return false if the string addresses are not equal
because two strings with different address may still have identical
contents.