Add a new "index-stats" tool.
This tool dumps statistics about the indexes in a pstore database
file. Currently the output is in CSV format to make it easy to drop
into a spreadsheet.

The tool is intended to enable me to evaluate the efficacy of
different hash functions to ensure that the resulting index trees are
reasonably well laid out.