Move maybe<> from adt/ back to support/.
The UTF-8 conversion code (in support/) uses maybe<> and it isn't
really possible to move this code out of support/ at the moment. I
can (sort of) justify maybe<> living in support even though it's
clearly an ADT because it's really a polyfill for std::optional<> in
C++17. We can probably remove maybe<> once its safe to move to that
newer language standard.