Correct PSTORE_ALWAYS_SPANNING so that 0 size requests never span.
When a writable "spanning" block is flushed to disk, storage::copy()
must compute the start address for the write. This must lie within the
allocated storage. If a request for 0 bytes is made, the resulting
start address could be at the start of an as-yet unallocated memory
region.

Previously, PSTORE_ALWAYS_SPANNING would mean that _every_
request (even those for 0 bytes) was spanning; if the debugging mode
is disabled, then 0 byte sizes are never spanning. This change mirrors
the latter behaviour if ALWAYS_SPANNING is enabled.