Chunked internal nodes (#60)
* Use chunked_vector to store in-heap internal-nodes.
Internal nodes are allocated using a chunked-vector. This allocates
memory in lumps sufficent for internal_nodes_per_chunk entries. This
is then consumed as new in-heap internal nodes are created.

The effect of this is to significantly reduce the number of memory
allocations performed by the index code under heavy insertion pressure
as thus give us a little extra performance. The cost is that we may
waste as much as one chunk's worth of memory [minus
sizeof (internal_node)].

* Use internal_node::size_bytes to compute the storage size.

* Assert that size()-1 is valid.