The begin() iterator was bogus after splice() on an empty CV.
The begin() function assumes that if the container is not empty then
the first chunk contains the first collection of members. Not
unreasonable... However, if splice() were used on an empty CV then it
did not consider the initial (empty) chunk. In this case, this chunk
must be replaced rather than retained.

Added a unit test for this case.