Add build_* to the list of files/directories in .gitignore.
I use the little python program at ./utils/make_build.py to quickly
run cmake with all of the proper arguments. By default, it creates a
directory named after the host OS (build_mac on macOS, build_win32 on
Windows, and so on). The change adds that pattern to the list that we
don't want under source control.