Fix for assertion when using the LLVM disassembler.
Commit SNSystems/llvm-project-prepo@0c369da upgraded the
llvm-project-prepo to the LLVM 11.x branch point. Unfortunately this
introduced a regression in pstore-dump where the call to
InitMCTargetOptionsFromFlags() triggered an assertion.

It seems likely that this call was always wrong: the "flags" that it's
talking about appear to be drawn from the LLVM command-line options,
which obviously isn't going to work with the pstore command-line
parser.

We now pass a default-initialized MCTargetOptions to createMCAsmInfo.

Add a system test to avoid this happening again. This means adding the
capability of running LLVM tools from inside the pstore test suite.