Add support for bash style comments.
Although comments aren't part of the JSON specification they are
useful. This commit adds optional support for comments starting with a
hash ('#') and extending to the rest of the line.

The use-case here is two-fold. First, the pstore-export utility can
document the requested string when emitting an index into the names
array. That helps with the readability of the exported data. Second,
it means that I can add comments containing llvm-lit instructions to
write tests for pstore and for the repository linker (rld).