Add the pstore-export and pstore-import tools (#70)
a new “exchange” library to pstore which is used to implement the pstore-export and pstore-import tools. The need for these tools is clear:

To enable a user to migrate information from one version of pstore to a newer version.
To trim a database back to a minimal set for fault investigation.
To enable LIT-style tests to be written for tools such as repo2obj and rld. (This latter use-case may be better resolved once llvm-project-prepo issue 4 is resolved although this may not happen immediately.)
The import/export feature set is extremely minimal ATM. The idea here is to get the core functionality committed to the master branch and refine from there. At the time of writing, the import process is likely somewhat slower than it needs to be because a lot of objects are dynamically allocated. This can improved easily enough later.

The export format currently uses JSON. We can investigate more efficient alternatives such as ProtoBuf in the future. The flavour of JSON in use includes the use of comments which are, of course, not permitted by a strictly conforming parser. (The change includes adding support for comments to the pstore parser.) These will become optional in future.

The library and tools are now at the point where they can be useful, I think. Apart from the brutally minimal feature set, there are a couple of limitations right now:

1. pstore-import can recreate a database, but not ticket files. These are needed for the third (LIT) use-case above.
2. It doesn’t yet im/export the write index as used by the pstore-read/pstore-write demo tools.
3. There is no support for “linked definition” sections ATM. There are compiler warnings about unused formal arguments that reflect this. See the comments here for more information.
4. The schema (pstore/utils/exchange_schema/pstore.schema.json) doesn't yet completely describe the format in use.