The database UUID header field is now ID.
The idea is that each database has an ID which identifies it and can
be used by external references (specifically, the Program Repository's
ticket files) to connect a digest and its owning database. Previously,
the database UUID field was largely unused. This allows a consumer to
report that a ticket file is being used with the wrong database file
rather than simply to report that a particular digest is now found in
the index which is much less helpful.