Change the get_sstring_view() interface to encourage correct use.
Whilst modifying the linked-definition code in the compiler
back-end, I noticed a probable bug in the usage of
get_sstring_view(). The code accessed the raw_sstring_view member of
the result and discarded the shared_sstring_view owner. The probable
bug turned out to be real: the owning memory was being discarded
resulting in a potential crash.

To reduce the probability of this happening again, I've changed the
interface so that the owning memory is passed by reference. This means
that the object's lifetime is likely to be greater. (It must be that
of the raw_sstring_view.)