Fix to install pstore header files.
When setting the PUBLIC_HEADER property of a target, we need to use
the full path of each of the header files. Fortunately, this has
already been produced: all we need to do is use ${include_files}
rather than ${arg_INCLUDES}.

Remove the explicit COMPONENT: this prevents header files from being
installed.