pstore-dump now reports the logical rather than physical file size.
I felt that reporting the physical size of the file wasn't
particularly useful: the information is easy to see in a directory
listing. Instead, the logical size of the file (which may be
considerably less is we're not in POSIX_SMALL_FILES mode) is a value
that you can't see any other way via pstore-dump.