Add a PSTORE_CLANG_TIDY_ENABLED option. Update the disabled checks.
Add a new cmake option PSTORE_CLANG_TIDY_ENABLED, which defaults to
Off, which allows user control over whether the clang-tidy targets are
generated. There were quite a few of them which bloated the list for
those cases where we're not interested in clang-tidy results.

Minor update to the list of disabled checks.