Elimninate string return of uint128::to_hex_string().
New version of the function [to_hex()] takes an output-iterator to
decouple the container in use from the function generating its
content. We know that a digest has a fixed length, so std::array<> can
be used instead.

write-digest is now a function common to all the places where a digest
is emitted.

Pre-reserve space in the names container.