Fix for #82.
When the user passes 0 as the port number to be used, we now pick a
free ephemeral port. The --announce-http-port displays a message when
the HTTP server is running and which port it is using.

Eliminated the use of a timeout in httpd.py in favour of waiting for
the output from --annount-http-port. The HTTP server is now disabled
using the new --disable-http switch rather than by requesting port 0.