Remove the --contents switch.
Unless the input database was really, really small, this switch would
produce a huge amount of output which ammounted to little more than a
hex dump of the file. This vast output meant that the --all switch was
a good as useless on larger inputs.

Correct the description of the --all switch.