Teach pstore-dump about x86_64 relocation names.
It's useful to be able to see the relocation names in pstore-dump's
output rather than just the raw type numbers. When built inside LLVM,
dump now knows how to dump relocation names for x86_64.

Start to rationalize the way that dump parameters are passed to the
make_value() implementations as a single 'parameters' struct rather
than just a random(ish) colleciton of extra function arguments.