Define an enum to reference strength.
The arguments used to build an external references have become rather
hard to read: passing a collection of numbers and now an extra bool
are all fairly meaningless.

Rather than passing a simple bool to the external_fixup ctor, we now
use an enum with explicit names for "weak" and "strong" references.