Fix for crash if parse_command_line_options() is called twice.
One of the forms of parse_command_line_options() implicitly added an
option for the '--help' switch. This object is on the function's
stack. Unfortunately, the option ctor adds the object to the global
options list. If this list was not reset before calling
parse_command_line_options() a second time, it would contain a
dangling pointer to that original '--help' instance and we crash.

To fix the problem, we now remove options from the global list in
their dtor.

The use of a global list remains far from ideal, though...