Add support for array and object trailing commas.
Trailing commas for arrays and objects are not permitted by the JSON
spec, but they are a useful convenience particularly when writing JSON
by hand. This is something that I'm doing fairly frequently as I write
tests for rld or repo2obj so it's something that I would like
pstore-import to support.