chunked_vector<> resize (#94)
I recently realized that rld had a bug in its handling of the case of a fragment shared by multiple definitions. If such a fragment has one or more external fixups which reference definitions with internal linkage then the linker must ensure that those fixups are resolved to the correct local definition.

Previously, symbol resolution was performed on each section’s fixups at most once. It must however be performed each time that a definition is added to the symbol table to ensure that external fixups always resolve to the correct internal definition. To fix this problem, I am now attaching the collection of referenced symbols to the each definition’s symbol-table entry. Previously “shadow memory” associated with the fragment was used for this purpose. 

The “collection of referenced symbols” is a contiguous sequence of rld::Symbol * (this is important for the performance of the code that applies the fixups during the linker’s copy phase). My intention is to use one CV per worker thread to store the symbol-resolution output without needing to add locks.For each section processed, the code will resize the container to guarantee a contiguous block of entries sufficient for that section’s external fixups.

All of this means that we need an implementation of resize() for CV.