Add comments to the names and paths arrays.
We already (optionally) produces comments that relate numbers in the
various structures to the name that they represent. This change
completes the circle by adding a comment for the numbers that
correspond to each name. This means that each string has a comment
giving its number and each number has a comment giving its string.

This is aimed at making the exported JSON a little easier for someone
to follow.