Resolve a warning from clang about run_broker having no prototype
Add a header file to contain the exports from brokerd/brokerd.cpp (as
was). Rename this TU as run_module since that's the name of the
function that it exports. run_broker() shouldn't be in the broker
library's namespace since it isn't part of the broker library. Include
file sorting.