Add database::getrou() functions.
A family of functions which return read-only unique_ptr<>-managed
memory in the store. These mirror the existing getro() functions which
return std::shared_ptr<>. The idea is that we avoid the costly atomic
reference counting operations of the shared_ptr<> versions when we
want the best posible performance, and don't need this sort of
memory-management help, esp. in rld.