Fix for a GCC 5.5.0 Ci build failure introduce by commit 329d12fe.
A workaround for an issue with some versions GNU libc++ that was
exposed by the GCC 5.5.0 CI/ build. Prior to a change in 2018-09-18,
std::is_default_constructible<> would return true for
pstore::unique_pointer<> but a static assertion would fire when the
default constructor was used.

The fix injects a specialization of BuiltinDefaultValue<> into Google
Mock which bypasses std::is_default_constructible<>.

Tweak the DbArchiveReadSpan fixture so that it avoids inheriting from
EmptyStore but instead aggregates InMemoryStore.